+++ internal/service/apikey/data_source_api_key.go
@@ -1,4 +1,4 @@
-package mongodbatlas
+package apikey
 
 import (
 	"context"
@@ -10,7 +10,7 @@ import (
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
 )
 
-func DataSourceAPIKey() *schema.Resource {
+func DataSource() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: dataSourceMongoDBAtlasAPIKeyRead,
 		Schema: map[string]*schema.Schema{
