+++ internal/service/apikey/resource_api_key_test.go
@@ -1,4 +1,4 @@
-package mongodbatlas_test
+package apikey_test
 
 import (
 	"context"
